chore: rename lib/gui/diagnostic => lib/gui/Diagnostic
