+++ src/apps/deskflow-gui/MainWindow.cpp
@@ -17,9 +17,9 @@
 #include "base/String.h"
 #include "common/Settings.h"
 #include "gui/Constants.h"
+#include "gui/Diagnostic.h"
 #include "gui/Logger.h"
 #include "gui/core/CoreProcess.h"
-#include "gui/diagnostic.h"
 #include "gui/ipc/DaemonIpcClient.h"
 #include "gui/messages.h"
 #include "gui/string_utils.h"
