+++ src/apps/deskflow-gui/main.cpp
@@ -9,8 +9,8 @@
 #include "MainWindow.h"
 #include "common/constants.h"
 #include "gui/Constants.h"
+#include "gui/Diagnostic.h"
 #include "gui/Logger.h"
-#include "gui/diagnostic.h"
 #include "gui/dotenv.h"
 #include "gui/messages.h"
 #include "gui/string_utils.h"
