+++ src/lib/gui/Diagnostic.h
